Move FAB session table creation to FAB provider
We need to create the `session` table in the provider db manager, not in
the core db utils.
